examples/tutorials/basic/odmantic/requirements.txt
starlette-admin
uvicorn
odmantic
pydantic[email]
